Refactor: ComparisonTable utility/helper functions refactored.
