+++ src/App.tsx
@@ -1,13 +1,67 @@
-import { useCallback, useMemo, useState } from "react";
-import { useQuery } from "react-query";
 import { Spin } from "antd";
+import { useQuery } from "react-query";
+import { useCallback, useState } from "react";
 
-import SearchBox from "./components/Searchbox/index";
-import ComparisonTable from "./components/ComparisonTable";
 import DownloadsChart from "./components/Chart";
+import SearchBox from "./components/Searchbox/index";
 import Recommendation from "./components/Recommendation";
+import ComparisonTable from "./components/ComparisonTable";
 
-import { DATA_KEYS, getTableData, parseNPMSResponse } from "./utils";
+import { parseNPMSResponse } from "./utils";
+import { TableDataDefinition } from "./types/types";
+
+export const DATA_KEYS: Array<TableDataDefinition> = [
+  {
+    key: "name",
+    label: "Package Name",
+    transform: (data: string) => data || "N/A",
+  },
+  {
+    key: "version",
+    label: "Version",
+    transform: (data: string) => data || "N/A",
+  },
+  {
+    key: "keywords",
+    label: "Keywords",
+    transform: (data: Array<string>) => {
+      if (!data) return "N/A";
+      if (data?.length > 10) {
+        return data.slice(0, 10).join(", ");
+      } else return data.join(", ");
+    },
+  },
+  {
+    key: "links",
+    label: "Repository",
+    transform: (data: { [x: string]: string }) => {
+      return Object.keys(data).map((key) => {
+        if (key === "npm") return null;
+        return (
+          <a target="_blank" key={key} href={data[key]}>
+            {key}
+          </a>
+        );
+      });
+    },
+  },
+  {
+    key: "license",
+    label: "License",
+    transform: (data: string) => data || "N/A",
+  },
+  {
+    key: "publisher",
+    label: "Authors/Publishers",
+    transform: (data: { username: string; email: string }) => data.email,
+  },
+  {
+    key: "maintainers",
+    label: "Maintainers",
+    transform: (data: Array<{ username: string; email: string }>) =>
+      data[0].email,
+  },
+];
 
 const fetchSelectedPackages = async (data: Array<string>) => {
   const response = await fetch("https://api.npms.io/v2/package/mget", {
@@ -37,8 +91,6 @@ function App() {
     setSelected(selectedArr);
   }, []);
 
-  const tableData = useMemo(() => getTableData(data), [data]);
-
   return (
     <div className="page">
       <h2>NPM Package Comparator</h2>
@@ -49,7 +101,7 @@ function App() {
         data && (
           <>
             <ComparisonTable
-              dataSource={tableData}
+              dataSource={parseNPMSResponse(data)}
               dataDefinition={DATA_KEYS}
             />
             <DownloadsChart data={parseNPMSResponse(data)} />
