+++ src/components/ComparisonTable/ComparisonTable.tsx
@@ -1,14 +1,16 @@
 import { FC } from "react";
 import { Col, Row } from "antd";
 
+import { ParsedResponse, TableProps } from "../../types/types";
+
 import "./styles.css";
-import { TableProps } from "../../types/types";
 
 const ComparisonTable: FC<TableProps> = ({ dataSource, dataDefinition }) => {
   if (!dataSource) return null;
   return (
     <div className="grid-table">
       {dataDefinition.map((item) => {
+        const key = item.key as keyof ParsedResponse;
         return (
           <Row className="row" key={item.key}>
             <Col className="column header" md={6} sm={4} lg={8} xl={4}>
@@ -22,7 +24,7 @@ const ComparisonTable: FC<TableProps> = ({ dataSource, dataDefinition }) => {
               lg={8}
               xl={10}
             >
-              {item.transform(dataSource[item.key]?.[0])}
+              {item.transform(dataSource[0]?.[key])}
             </Col>
             <Col
               className={"column " + item.key}
@@ -32,7 +34,7 @@ const ComparisonTable: FC<TableProps> = ({ dataSource, dataDefinition }) => {
               lg={8}
               xl={10}
             >
-              {item.transform(dataSource[item.key]?.[1])}
+              {item.transform(dataSource[1]?.[key])}
             </Col>
           </Row>
         );
